refactor(workflows/code-editor): split python providers into per-concern modules
The single 1255-line registerPythonProviders.ts mixed completion, hover,
signature-help, formatter, validator and code-actions in one file — any
tweak risked touching multiple editor behaviours.

Split into per-provider modules under `monaco/python/`:
  shared.ts        — types, regexes, marker codes, helpers (scanImports,
                     literal classifiers, default-value lookup, scaffolds)
  completion.ts    — registerCompletion + memberCompletion + itemKind
  hover.ts         — registerHover
  signatureHelp.ts — registerSignatureHelp
  formatter.ts     — registerFormatter
  validator.ts     — registerValidator + ValidatorHandle
  codeActions.ts   — registerCodeActions

`registerPythonProviders.ts` is now a thin orchestrator (~50 lines) that
wires the disposables together and forwards `setContract` to the
validator's revalidate hook.

CodeEditorModal imports the public types directly from `./monaco/python/shared`
— no re-export shim.
